Split e2e tests into seperate files
